requirements.txt
RPi.GPIO
adafruit-circuitpython-servokit
adafruit-circuitpython-motorkit
aiohttp
opencv-python
python-dotenv
